Rename and Move IMiunieDiscord
